+++ DRN.Framework.EntityFramework/Context/DrnContextDevelopmentConnection.cs
@@ -1,3 +1,4 @@
+using DRN.Framework.SharedKernel;
 using DRN.Framework.Utils.Settings;
 using Microsoft.Extensions.Configuration;
 
@@ -5,14 +6,6 @@ namespace DRN.Framework.EntityFramework.Context;
 
 public static class DrnContextDevelopmentConnection
 {
-    public const string PostgresDevelopmentPasswordKey = "PostgresDevelopmentPassword";
-    public const string PostgresDevelopmentHostKey = "PostgresDevelopmentHost";
-    public const string PostgresDevelopmentPortKey = "PostgresDevelopmentPort";
-    public const string PostgresDevelopmentUsernameKey = "PostgresDevelopmentUsername";
-    public const string PostgresDevelopmentDatabaseKey = "PostgresDevelopmentDatabase";
-    public const string DefaultUsername = "postgres";
-    public const string DefaultDatabase = "drnDb";
-
     public static string GetConnectionString(IAppSettings appSettings, string name)
     {
         var connectionString = string.Empty;
@@ -20,13 +13,14 @@ public static class DrnContextDevelopmentConnection
             connectionString = devConnectionString;
         else
         {
-            var host = appSettings.Configuration.GetValue(PostgresDevelopmentHostKey, "postgresql");
-            var port = appSettings.Configuration.GetValue(PostgresDevelopmentPortKey, 5432);
-            var username = appSettings.Configuration.GetValue<string>(PostgresDevelopmentUsernameKey, DefaultUsername);
-            var database = appSettings.Configuration.GetValue<string>(PostgresDevelopmentDatabaseKey, DefaultDatabase);
-            var password = appSettings.Configuration.GetValue<string>(PostgresDevelopmentPasswordKey);
+            var host = appSettings.Configuration.GetValue(DbContextConventions.DevHostKey, "postgresql");
+            var port = appSettings.Configuration.GetValue(DbContextConventions.DevPortKey, 5432);
+            var username = appSettings.Configuration.GetValue<string>(DbContextConventions.DevUsernameKey, DbContextConventions.DefaultUsername);
+            var database = appSettings.Configuration.GetValue<string>(DbContextConventions.DevDatabaseKey, DbContextConventions.DefaultDatabase);
+            var password = appSettings.Configuration.GetValue<string>(DbContextConventions.DevPasswordKey);
+
             if (password != null)
-                connectionString = $"Host={host};Port={port};Database={database};User ID={username};password={password};";
+                connectionString = $"Host={host};Port={port};Database={database};User ID={username};password={password};Multiplexing=true;Max Auto Prepare=200;Maximum Pool Size=20;Application Name={AppConstants.ApplicationName};";
         }
 
         ArgumentException.ThrowIfNullOrWhiteSpace(connectionString);