+++ DRN.Framework.Utils/DependencyInjection/Attributes/HasServiceCollectionModuleAttribute.cs
@@ -14,7 +14,8 @@ public abstract class HasServiceCollectionModuleAttribute : Attribute
     public static HasServiceCollectionModuleAttribute GetModuleAttribute(Type type) =>
         (HasServiceCollectionModuleAttribute)type.GetCustomAttributes().Single(a => a.GetType().IsAssignableTo(typeof(HasServiceCollectionModuleAttribute)));
 
-    public virtual void PostStartupValidation(object service, IServiceProvider serviceProvider)
+    public virtual async Task PostStartupValidationAsync(object service, IServiceProvider serviceProvider)
     {
+        await Task.CompletedTask;
     }
 }